pyproject.toml
[project]
name = "MikuChatSDK"
version = "0.2.0"
description = "Python 3 SDK for MikuChat. Check https://doc.yuuz12.top/web/#/5/ "
authors = [
    {name = "Number_Sir", email = "number_sir@126.com"},
]
maintainers = [
    {name = "Number_Sir", email = "number_sir@126.com"},
]
dependencies = [
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
keywords = ["yuuz12", "shiruku", "mikuchat"]

[project.urls]
documentation = "https://doc.yuuz12.top/web/#/5/58"
repository = "https://github.com/NumberSir/mikuchat-sdk"
issues = "https://github.com/NumberSir/mikuchat-sdk/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"